test: centralize Lean vocabulary helper ownership
