+++ crates/gents-desktop-bridge/src/snapshot/tests/session_state.rs
@@ -1,6 +1,6 @@
 use super::*;
 
-#[path = "../../../../../crates/gents/src/lean_vocab_test.rs"]
+#[path = "../../../../../crates/gents/src/lean_vocab_test/support.rs"]
 mod lean_vocab_test;
 
 use gents::llm::message::{
