+++ crates/gents-desktop-core/src/client/mutations/chat/request.rs
@@ -818,7 +818,7 @@ async fn fetch_retry_root(node: &EmbeddedNode, request_id: &str) -> Result<Optio
 }
 
 #[cfg(test)]
-#[path = "../../../../../gents/src/lean_vocab_test.rs"]
+#[path = "../../../../../gents/src/lean_vocab_test/support.rs"]
 mod lean_vocab_test;
 
 #[cfg(test)]
